requirements.txt
fake_useragent~=0.0.9
matplotlib~=1.5.1
requests~=2.9.1
python-dateutil~=2.6.0
